+++ fabric-1.21.4/src/main/java/net/william278/cloplib/events/RaidStarted.java
@@ -17,23 +17,24 @@
  *  limitations under the License.
  */
 
-package net.william278.cloplib.listener.events;
+package net.william278.cloplib.events;
 
 import net.fabricmc.fabric.api.event.Event;
 import net.fabricmc.fabric.api.event.EventFactory;
+import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.util.ActionResult;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import org.jetbrains.annotations.NotNull;
 
-public final class FluidFlowsHorizontally {
+public final class RaidStarted {
 
     @NotNull
-    public static final Event<BeforeFluidFlows> EVENT = EventFactory.createArrayBacked(
-            BeforeFluidFlows.class,
-            (callbacks) -> (world, from, to) -> {
-                for (BeforeFluidFlows listener : callbacks) {
-                    final ActionResult result = listener.flow(world, from, to);
+    public static final Event<BeforeRaidStarted> EVENT = EventFactory.createArrayBacked(
+            BeforeRaidStarted.class,
+            (callbacks) -> (world, pos, player) -> {
+                for (BeforeRaidStarted listener : callbacks) {
+                    final ActionResult result = listener.started(world, pos, player);
                     if (result != ActionResult.PASS) {
                         return result;
                     }
@@ -44,10 +45,10 @@ public final class FluidFlowsHorizontally {
     );
 
     @FunctionalInterface
-    public interface BeforeFluidFlows {
+    public interface BeforeRaidStarted {
 
         @NotNull
-        ActionResult flow(World world, BlockPos from, BlockPos to);
+        ActionResult started(World world, BlockPos from, PlayerEntity player);
 
     }
