+++ fabric-1.21.4/src/main/java/net/william278/cloplib/listener/FabricUseEntityListener.java
@@ -1,0 +1,4 @@
+package net.william278.cloplib.listener;
+
+public interface FabricUseEntityListener extends FabricListener {
+}
