+++ fabric-1.21.4/src/main/resources/cloplib.mixins.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "net.william278.cloplib.listener.mixins",
+  "package": "net.william278.cloplib.mixins",
   "compatibilityLevel": "JAVA_17",
   "server": [
     "AbstractBlockStateMixin",
